+++ core/domain/interface/delivery/area_value.go
@@ -10,7 +10,7 @@ package delivery
 
 //中国省市行政规划
 type AreaValue struct {
-	Id   int64  `db:"id" pk:"yes" auto:"no"`
-	Pid  int64  `db:"pid"`
+	Id   int32  `db:"id" pk:"yes" auto:"no"`
+	Pid  int32  `db:"pid"`
 	Name string `db:"name"`
 }
