+++ core/domain/interface/member/invitation_manager.go
@@ -12,7 +12,7 @@ import "go2o/core/dto"
 
 type IInvitationManager interface {
 	// 判断是否由会员邀请
-	InvitationBy(memberId int64) bool
+	InvitationBy(memberId int32) bool
 
 	// 获取我邀请的会员
 	GetInvitationMembers(begin, end int) (total int, rows []*dto.InvitationMember)
