+++ core/domain/interface/merchant/level_manager.go
@@ -10,12 +10,12 @@ package merchant
 
 type (
 	MemberLevel struct {
-		Id         int64 `db:"id" auto:"yes" pk:"yes"`
-		MerchantId int64 `db:"merchant_id"`
+		Id         int32 `db:"id" auto:"yes" pk:"yes"`
+		MerchantId int32 `db:"merchant_id"`
 		// 等级值(1,2,4,8,16)
-		Value      int64  `db:"value" `
+		Value      int32  `db:"value" `
 		Name       string `db:"name"`
-		RequireExp int64  `db:"require_exp"`
+		RequireExp int32  `db:"require_exp"`
 		Enabled    int    `db:"enabled"`
 	}
 
@@ -25,22 +25,22 @@ type (
 		GetLevelSet() []*MemberLevel
 
 		// 获取等级
-		GetLevelById(id int64) *MemberLevel
+		GetLevelById(id int32) *MemberLevel
 
 		// 根据等级值获取等级
-		GetLevelByValue(value int64) *MemberLevel
+		GetLevelByValue(value int32) *MemberLevel
 
 		// 获取下一个等级
-		GetNextLevel(value int64) *MemberLevel
+		GetNextLevel(value int32) *MemberLevel
 
 		// 删除等级
-		DeleteLevel(id int64) error
+		DeleteLevel(id int32) error
 
 		// 保存等级
-		SaveLevel(*MemberLevel) (int64, error)
+		SaveLevel(*MemberLevel) (int32, error)
 
 		// 根据经验值获取等级值
-		GetLevelValueByExp(exp int64) int64
+		GetLevelValueByExp(exp int32) int32
 
 		// 初始化默认等级
 		InitDefaultLevels() error
